restructure lib and move everything into src
